requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
pandas>=2.0.0
google-generativeai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
webdriver-manager>=4.0.0